+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ---
 
-## [Unreleased] 
+## [0.22.0] - 2026-03-26 @ 18:14 - Tweaked Layout Release 
 
 ### Added
 - Complete network configuration is now visible in the TUI. The network configuration now shows: